Rename values to exprs
